pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gpie"
version = "0.2.2"
description = "Graph-based Probabilistic Inference Engine"
authors = [{ name="Hajime Ueda", email="nd.hajimeueda@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"

# ---- Core dependencies ----
dependencies = [
    "numpy>=2.2.6",
    "networkx>=3.3",
]

# ---- Optional dependencies ----
[project.optional-dependencies]
gpu = [
    # for GPU acceleration
    "cupy-cuda12x>=13.4.2",
]
fft = [
    # for optimized FFT on CPU
    "pyfftw>=0.15",
]
viz = [
    # for visualization
    "matplotlib>=3.9.0",
    "pygraphviz>=1.10",
    "bokeh>=3.7",
]
dev = [
    # development and testing
    "pytest",
    "pytest-cov",
    "coverage",
    "tqdm",
    "scikit-image>=0.25",
]

[tool.setuptools]
packages = ["gpie"]
